+++ kitty/shaders.c
@@ -942,7 +942,7 @@ draw_scrollbar(const UIRenderData *ui) {
     if (!window || !screen || !has_scrollbar(window, screen)) return;
 
     color_type bar_color = scrollbar_color(screen, OPT(scrollbar_handle_color)), track_color = scrollbar_color(screen, OPT(scrollbar_track_color));
-    double cell_frac = screen->pixel_scroll_offset_y / screen->cell_size.height;
+    double cell_frac = (double)screen->pixel_scroll_offset_y / screen->cell_size.height;
     if (!OPT(pixel_scroll)) cell_frac = 0;
     float bar_frac = (float)(screen->scrolled_by + cell_frac) / MAX(1u, (float)screen->historybuf->count);
     float opacity = OPT(scrollbar_handle_opacity);
